Leaves only spawn inside tree

+++ COMP220/COMP220_Portfolio/ParticleEffectManager.h
@@ -46,7 +46,7 @@ private:
 	//! int for the maximum number of particles that can exist at one time
 	int maxParticleNumber = 20;
 	//! float starting point for particle
-	float intialYposition = 0;
+	int intialYposition = 0;
 	//! the size f the particle when rendered.
 	float particleSize = 0.1;
 };
